+++ build-module/components/collaborative-editing/use-yjs/yjs-doc.js
@@ -42,7 +42,9 @@ export function createDocument(_ref) {
     relativePositionManager.peers.setAbsolutePositions(doc); // Change received from peer, or triggered by self undo/redo
 
     if (origin !== identity) {
-      const newData = postDocToObject(doc);
+      const newData = postDocToObject(doc, {
+        sanitize: true
+      });
       yDocTriggeredChangeListeners.forEach(listener => listener(newData));
       relativePositionManager.self.setAbsolutePosition(doc);
     }